+++ Apermo/Tests/WordPress/NoAdminAjaxUnitTest.php
@@ -1,0 +1,44 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * Unit test for the NoAdminAjax sniff.
+ *
+ * @package Apermo\Tests\WordPress
+ */
+
+namespace Apermo\Tests\WordPress;
+
+use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
+
+/**
+ * Unit test for Apermo.WordPress.NoAdminAjax.
+ */
+class NoAdminAjaxUnitTest extends AbstractSniffUnitTest {
+
+	/**
+	 * Returns the lines where errors should occur.
+	 *
+	 * @param string $testFile The name of the test file being tested.
+	 *
+	 * @return array<int, int>
+	 */
+	protected function getErrorList( $testFile = '' ) {
+		return [
+			7  => 1, // wp_ajax_ single quotes.
+			10 => 1, // wp_ajax_nopriv_ single quotes.
+			13 => 1, // wp_ajax_ double quotes.
+		];
+	}
+
+	/**
+	 * Returns the lines where warnings should occur.
+	 *
+	 * @param string $testFile The name of the test file being tested.
+	 *
+	 * @return array<int, int>
+	 */
+	protected function getWarningList( $testFile = '' ) {
+		return [];
+	}
+}